book search is glitching
